HTTPC-173: Extract URI methods to a utility class (#564)
(cherry picked from commit f2fb6eb406e64a8388e877e61316e680e264f998)
