+++ src/main/java/org/mule/extension/http/internal/request/HttpRequestOperations.java
@@ -12,8 +12,9 @@ import static org.mule.extension.http.internal.HttpConnectorConstants.CONNECTOR_
 import static org.mule.extension.http.internal.HttpConnectorConstants.HTTP_ENABLE_PROFILING;
 import static org.mule.extension.http.internal.HttpConnectorConstants.REQUEST;
 import static org.mule.extension.http.internal.HttpConnectorConstants.RESPONSE;
+import static org.mule.extension.http.internal.request.UriUtils.buildPath;
+import static org.mule.extension.http.internal.request.UriUtils.resolveUri;
 import static org.mule.runtime.extension.api.annotation.param.MediaType.ANY;
-import static org.mule.runtime.http.api.utils.HttpEncoderDecoderUtils.encodeSpaces;
 
 import org.mule.extension.http.api.HttpResponseAttributes;
 import org.mule.extension.http.api.error.HttpErrorMessageGenerator;
@@ -23,8 +24,8 @@ import org.mule.extension.http.api.request.validator.ResponseValidator;
 import org.mule.extension.http.api.request.validator.SuccessStatusCodeValidator;
 import org.mule.extension.http.internal.HttpMetadataResolver;
 import org.mule.extension.http.internal.request.client.HttpExtensionClient;
-import org.mule.extension.http.internal.request.profiling.HttpRequestResponseProfilingDataProducerAdaptor;
 import org.mule.extension.http.internal.request.profiling.HttpProfilingServiceAdaptor;
+import org.mule.extension.http.internal.request.profiling.HttpRequestResponseProfilingDataProducerAdaptor;
 import org.mule.runtime.api.exception.DefaultMuleException;
 import org.mule.runtime.api.exception.MuleException;
 import org.mule.runtime.api.lifecycle.Disposable;
@@ -55,17 +56,14 @@ import org.mule.runtime.extension.api.runtime.operation.Result;
 import org.mule.runtime.extension.api.runtime.parameter.CorrelationInfo;
 import org.mule.runtime.extension.api.runtime.process.CompletionCallback;
 import org.mule.runtime.extension.api.runtime.streaming.StreamingHelper;
-import org.mule.runtime.http.api.HttpConstants;
-
-import java.io.InputStream;
-import java.util.HashMap;
-import java.util.List;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import javax.inject.Inject;
 import javax.inject.Named;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import java.io.InputStream;
+import java.util.HashMap;
+import java.util.List;
 
 public class HttpRequestOperations implements Initialisable, Disposable {
 
@@ -176,36 +174,12 @@ public class HttpRequestOperations implements Initialisable, Disposable {
     }
   }
 
-  private String resolveUri(HttpConstants.Protocol scheme, String host, Integer port, String path) {
-    // Encode spaces to generate a valid HTTP request.
-    return scheme.getScheme() + "://" + host + ":" + port + encodeSpaces(path);
-  }
-
   private int resolveResponseTimeout(Integer responseTimeout) {
     if (muleContext.getConfiguration().isDisableTimeouts()) {
       return WAIT_FOR_EVER;
     } else {
       return responseTimeout != null ? responseTimeout : muleContext.getConfiguration().getDefaultResponseTimeout();
     }
-  }
-
-  protected String buildPath(String basePath, String path) {
-    String resolvedBasePath = basePath;
-    String resolvedRequestPath = path;
-
-    if (!resolvedBasePath.startsWith("/")) {
-      resolvedBasePath = "/" + resolvedBasePath;
-    }
-
-    if (resolvedBasePath.endsWith("/") && resolvedRequestPath.startsWith("/")) {
-      resolvedBasePath = resolvedBasePath.substring(0, resolvedBasePath.length() - 1);
-    }
-
-    if (!resolvedBasePath.endsWith("/") && !resolvedRequestPath.startsWith("/") && !resolvedRequestPath.isEmpty()) {
-      resolvedBasePath += "/";
-    }
-
-    return resolvedBasePath + resolvedRequestPath;
   }
 
   @Override
